refactor: улучшить работу кнопки moreMovies
