+++ src/components/MoviesCardList/MoviesCardList.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react";
+// import { useState, useEffect } from "react";
 import MoviesCard from "../MoviesCard/MoviesCard";
 import Preloader from "../Preloader/Preloader";
 
@@ -8,38 +8,41 @@ function MoviesCardList({
   saved,
   savedMovies,
   onMovieButtonCkick,
+
+  moreMoviesButton,
+  handleClickMoreMovies
 }) {
-  const [moreMovies, setMoremuvies] = useState(2);
-  const [moviesAmount, setMoviesAmount] = useState(5);
+  // const [moreMovies, setMoremuvies] = useState(2);
+  // const [moviesAmount, setMoviesAmount] = useState(5);
 
-  useEffect(() => {
-    window.addEventListener("resize", resize);
-    resize();
-    // setTimeout(() => {
-    //   window.addEventListener("resize", resize);
-    //   resize();
-    // }, 100);
-    return () => window.removeEventListener("resize", resize);
-  }, []);
+  // useEffect(() => {
+  //   window.addEventListener("resize", resize);
+  //   resize();
+  //   // setTimeout(() => {
+  //   //   window.addEventListener("resize", resize);
+  //   //   resize();
+  //   // }, 100);
+  //   return () => window.removeEventListener("resize", resize);
+  // }, []);
 
-  function resize() {
-    if (window.innerWidth > 1025) {
-      setMoremuvies(4);
-      setMoviesAmount(16);
-    } else if (window.innerWidth > 768) {
-      setMoremuvies(3);
-      setMoviesAmount(9);
-    } else if (window.innerWidth > 500) {
-      setMoremuvies(2);
-      setMoviesAmount(8);
-    } else {
-      setMoviesAmount(5);
-    }
-  }
+  // function resize() {
+  //   if (window.innerWidth > 1025) {
+  //     setMoremuvies(4);
+  //     setMoviesAmount(16);
+  //   } else if (window.innerWidth > 768) {
+  //     setMoremuvies(3);
+  //     setMoviesAmount(9);
+  //   } else if (window.innerWidth > 500) {
+  //     setMoremuvies(2);
+  //     setMoviesAmount(8);
+  //   } else {
+  //     setMoviesAmount(5);
+  //   }
+  // }
 
-  function handleClickMoreMovies() {
-    setMoviesAmount(moviesAmount + moreMovies);
-  }
+  // function handleClickMoreMovies() {
+  //   setMoviesAmount(moviesAmount + moreMovies);
+  // }
 
   function isMovieSaved(movie) {
     return savedMovies.some((savedMovie) => savedMovie.movieId === movie.id);
@@ -51,7 +54,8 @@ function MoviesCardList({
         <Preloader />
       ) : (
         <div className="cards__list">
-          {movies.slice(0, moviesAmount).map((data, i) => (
+          {/* {movies.slice(0, moviesAmount).map((data, i) => ( */}
+          {movies.map((data, i) => (
             <MoviesCard
               movie={data}
               key={i}
@@ -61,10 +65,13 @@ function MoviesCardList({
           ))}
         </div>
       )}
-      {moviesAmount < movies.length && (
+      {/* {moviesAmount < movies.length && ( */}
+      {moreMoviesButton ? (
         <button className="cards__button" onClick={handleClickMoreMovies}>
           Ещё
         </button>
+      ) : (
+        <></>
       )}
     </section>
   );
